requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
djangorestframework==3.15.2
flower==2.0.1
humanize==4.11.0
kombu==5.4.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
